move easy rsa url to config
