+++ services/TS29222_CAPIF_API_Provider_Management_API/api_provider_management/db/db.py
@@ -6,7 +6,7 @@ from bson.codec_options import CodecOptions
 class MongoDatabse():
 
     def __init__(self):
-        self.config = Config().getConfig()
+        self.config = Config().get_config()
         self.db = self.__connect()
         self.provider_enrolment_details = self.config['mongo']['col']
         self.capif_users = self.config['mongo']['capif_users']
