+++ services/TS29222_CAPIF_API_Provider_Management_API/config.yaml
@@ -8,3 +8,9 @@ mongo: {
   'host': 'mongo',
   'port': "27017"
 }
+
+ca_factory: {
+  "url": "easy-rsa",
+  "port": "8083"
+}
+
