+++ services/register/config.yaml
@@ -6,3 +6,7 @@ mongo: {
   'host': 'mongo_register',
   'port': '27017'
 }
+ca_factory: {
+  "url": "easy-rsa",
+  "port": "8083"
+}
