+++ services/register/register_service/config.py
@@ -15,6 +15,6 @@ class Config:
 			self.my_config = yaml.safe_load(f)
 			f.close()
 
-	def getConfig(self):
+	def get_config(self):
 		return self.my_config
