+++ services/register/register_service/db/db.py
@@ -1,11 +1,11 @@
 from pymongo import MongoClient
-from .config import Config
+from ..config import Config
 
 
 class MongoDatabse():
 
     def __init__(self):
-        self.config = Config().getConfig()
+        self.config = Config().get_config()
         self.db = self.__connect()
         self.capif_users = self.config['mongo']['col']
